Fix #3323: Revert back to standard share sheet elements based on URL. (#5053)

+++ Client/Frontend/Browser/TemporaryDocument.swift
@@ -38,6 +38,8 @@ class TemporaryDocument: NSObject {
         }
     }
 
+    /// Get a deferred url to the locally stored file.
+    /// - Note: If an error occurs, the original web url will be returned
     func getURL() -> Deferred<URL> {
         if let url = localFileURL {
             let result = Deferred<URL>()
@@ -63,7 +65,8 @@ class TemporaryDocument: NSObject {
 
         return result
     }
-    
+
+    /// A callback available when there is a download response
     func onDocumentDownloaded(document: DownloadedResourceResponse?, error: Error?) {
         // Store the blob/data in a local temporary file.
         if let document = document, let data = document.data, !data.isEmpty {
@@ -72,7 +75,12 @@ class TemporaryDocument: NSObject {
             
             do {
                 try FileManager.default.createDirectory(at: tempDirectory, withIntermediateDirectories: true, attributes: nil)
-                try FileManager.default.removeItem(at: url)
+
+                // Delete the file if there is already one. We will replace it with a new file
+                if FileManager.default.fileExists(atPath: url.absoluteString) {
+                    try FileManager.default.removeItem(at: url)
+                }
+
                 try data.write(to: url, options: [.atomic])
                 
                 localFileURL = url
